+++ internal/model/user.go
@@ -2,5 +2,5 @@ package model
 
 type User struct {
 	ID      int
-	Balance int
+	Balance float64
 }
